requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
mysql==0.0.3
mysqlclient==2.1.1
PyMySQL==1.0.2
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.7
sqlparse==0.4.3
texttable==1.6.7
mysql-connector-python==8.0.32
